requirements.txt
flask
boto3
flasgger
flask
flask-swagger-ui
python-dotenv